# Make the number of fingerprint mutexes configurable