@beorn7 2016-06-02
Benchmark:

OLD:
BenchmarkFingerprintLockerParallel-4     2000000               697 ns/op
BenchmarkFingerprintLockerSerial-4      50000000                38.7 ns/op

NEW:
BenchmarkFingerprintLockerParallel-4     2000000               694 ns/op
BenchmarkFingerprintLockerSerial-4      50000000                40.2 ns/op