@fabxc 2016-06-02
👍 for configurability

Given how much trouble we had in the past with fingerprints, I'm worried about any changes to them unless absolutely necessary.