@beorn7 2016-06-02
Fingerprints are not changed. What's changing is the way we assign a preallocated mutex to a fingerprint. When the code was written, we naively assumed an "almost random" hash. fnv64a is all but that. Relying only on a change in a few least significant bits is needlessly bad.