@fabxc 2016-06-02
I realize that. Regardless, we are dealing with fingerprints here in a
critical area. Hence the question: do we really need it or did you just add
it along the way?

On Thu, Jun 2, 2016 at 11:41 PM Björn Rabenstein <notifications@github.com>
wrote:

> Fingerprints are not changed. What's changing is the way we assign a
> preallocated mutex to a fingerprint. When the code was written, we naively
> assumed an "almost random" hash. fnv64a is all but that. Relying only on a
> change in a few least significant bits is needlessly bad.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/prometheus/prometheus/pull/1701#issuecomment-223431274>,
> or mute the thread
> <https://github.com/notifications/unsubscribe/AEuA8kHCJ4sseJrvBVAyoudgaThYoYctks5qH031gaJpZM4IsyW6>
> .
>