@beorn7 2016-06-02
In the concrete case I'm currently working on, it only helps marginally (but so does the increase of mutexes – that's a different story I have to raise separately).
But the degenerate case of fingerprints that only change in a few bits is well known. Relying on just the least significant bits might bite us hard in subtle ways.
What exactly is critical about this change? The only critical requirement is to always assign the same mutex to the same fingerprint. This change can never fail on that.
You are referring to previous situations that were not well understood. Is there anything not analytically clear with this change?