@grobie 2016-06-02
What concerns me in general is the number of growing flags. From all projects I've heard of providing many configuration options, it eventually became very hard to get them right. Up to the point where people started to apply machine learning algorithms to figure out the optimal settings of a well-known VM for any given environment.

The Go virtual machine plays at the very other end of the spectrum, even GOMAXPROCS is rarely touched these days. And I image with garbage collection one could justify a large number of configuration options as well. In my understanding operational simplicity is a much wanted feature and also a goal of Prometheus.

How many people really need to have such configuration option?