@brian-brazil 2016-06-02
It seems natural to me that flags would grow over time, and this is one of many things in the codebase that should have a flag.

We can't automatically choose correct values for many constants, so we can either not expose them to users and have them run into performance/reliability issues; or expose them and allow power users configure them.