@bitfehler 2016-06-03
Just as my two cents here, without knowing much about your design principles or anything: I think having a flag is never a bad idea as long as the default works reasonably well for the vast majority of people. Hence, noone needs to bother with it, but if you actually do run into problems, if you dig deep enough, there is a knob to turn. 
Disclaimer: I might be biased due to the fact of being part of the minority that is having problems ;)