@fabxc 2016-06-03
As mentioned often, I totally agree with what @grobie is saying here. Go is a perfect example that you can actually go flag-less – compare it to JVM's options around GC and yet Go seems to work extremely well.

Regardless, having the flags only causes moderate code complexity, so it's fine. I just think we should have a regular `--help` output and an extended one so the regular user doesn't have to dig through 100 flags to find how to tweak the log level.