@fabxc 2016-06-03
As for the FP thing @beorn7 – it's easy to say "this last time we missed something but we don't do this time". I'm sure that was said before every hidden pitfall ever.

If you are confident that it won't break and has a performance relevant tweak, go for it though.
The benchmark seems to not show any improvement though. Or am I missing something?