@beorn7 2016-06-03
The benchmark is not using fingerprints that don't change in their least significant bits.
We could craft such a benchmark.

My main argument is that the code was written under the assumption that the fingerprints are pseudo random. By now we have understood that they aren't, and it is a quite common effect that two fingerprints for similar metrics vary only in a few bits, which could very well be somewhere else than at the end. If you have a setup where you scrape a massive amount of very similar metrics, they might all end up using the same mutex.

Or in other word: Had we had a better understanding of our fingerprinting algorithm, we would have written the code in the way I did now from the beginning. It's a _conceptual_ improvement, even without proven performance benefit. (And BTW, I still have to update the doc comment that explains why we have used only the least significant bits originally. Commit on its way…)

Now you are totally right that we have to weigh an actual performance benefit vs. the risk of a regression. _Just_ a conceptual improvement is not sufficient. We could invest a lot of our collective time to prove and discuss the performance part. What I'm trying here is to save our collective time by discussing the other side of the equation: the possible risk. I argue that applying a deterministic function consistently to a hash before picking the lock cannot cause a bug. (It could make the selection even less equally distributed, if the function was chosen unwisely, but that would only make things slower, not break anything.) Additionally, this code is excessively covered by our tests. I ran the concurrent test with the race detector to be super sure.