@beorn7 2016-06-03
WRT flag proliferation: It's great if a problem can be solved auto-magically. Tweaking a Prometheus server is a hard problem. I see myself tweaking servers in various different ways so that no common pattern has arrived yet. End goal would be to just provide flags like "I want the server to utilize 64GiB of RAM" or "I want the server to not exceed 1TiB of disk usage" and everything else will auto-tune. We are not there yet, and this PR is an excellent example for a case where we tried to avoid the flag but then bumped into reality and had to provide it after all.