@beorn7 2016-06-03
The promised doc change is in https://github.com/prometheus/docs/commit/c87d331ead3205e6946cc9b56b78de710e4588f6
I accidentally pushed it to the `next-release` branch directly.