@fabxc 2016-06-03
Okay, all good 👍 
Same on the flag – as said, I don't mind the flags so much but their impact on ease of use. Which we can solve via structuring `--help` better in the future.